Beta version 0.3.6! / Версия Beta 0.3.6! - added a lot of minerals; - changed the price for the construction of the plant; ------------------------------------------------------ - добавил много полезных ископаемых; - изменил цену для постройки завода;

+++ Assets/Scripts/Menus/PlanetMenu.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 using UnityEngine.Events;
-using UnityEngine.UI;
 
 namespace PlanetsColony
 {
@@ -18,7 +16,7 @@ namespace PlanetsColony
         private Camera _mainCamera = null;
 
         // временные переменные здесь:
-        private ulong _tempNeedMoney = 0;
+        private BigInteger _tempNeedMoney = 0;
 
         private void Awake()
         {
